temp : 원격지에서의 작업을 위해 미완성 기능 커밋

+++ src/main/java/com/hanghae/ecommerce/domain/order/item/OrderItem.java
@@ -31,7 +31,7 @@ public class OrderItem extends AbstractEntity {
 	@JoinColumn(name = "product_id", nullable = false)
 	private Product product;
 
-	private int quantity;
-	private int price;
+	private Long quantity;
+	private Long price;
 
 }
